chore: migrate `addScriptTag` handler
